% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byframe_corr.R
\name{byframe_corr}
\alias{byframe_corr}
\title{corrections for each frame}
\usage{
byframe_corr(framex, sdf, ProcessedIMU)
}
\arguments{
\item{framex}{frame of interest}

\item{sdf}{spectral data frame output from ortho fun}

\item{ProcessedIMU}{output from ortho fun}
}
\description{
This function places the center of each frame in the correct location
}
