% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degtorad.R
\name{degtorad}
\alias{degtorad}
\title{Degrees to radians conversion}
\usage{
degtorad(x)
}
\arguments{
\item{x}{Angle}
}
\description{
This function converts angles to radians for use in calculations
}
\examples{
degtorad(90)
}
\keyword{conversion}
