% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imu_demcombine.R
\name{imu_demcombine}
\alias{imu_demcombine}
\title{Combine DEM file with IMU}
\usage{
imu_demcombine(imu = imu.framematch,
  demfilelocation = "~/Ortho_Proc/DEM 1m RESCUE/dem_1m_m.bil",
  latMin = latMinIMU, lonMin = lonMinIMU)
}
\arguments{
\item{imu}{processed imu/frame file}

\item{demfile}{name of dem that is within local subfolder}

\item{latmin}{latitude of min elevation in imu - typically takeoff point}

\item{lonmin}{longitude of min elevation in imu - typically takeoff point}
}
\description{
This function combines the imu/frame file with dem to calculate accurate height throughout the flight
}
\examples{
imu_demcombine()
}
