% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imu_framecombine.R
\name{imu_framecombine}
\alias{imu_framecombine}
\title{IMU frame combine}
\usage{
imu_framecombine(folderloc = paste0(RemoteSenDataLoc, FolderLoc),
  imu = "imu_gps.txt")
}
\arguments{
\item{folderloc}{where are the frame files located}

\item{imu}{file name for imu_gps file, including extension .txt}
}
\description{
This function combines frame numbers with timestamps and calculates min IMU and location of min IMU to later match with DEM and correct for spatial variation in ground level.
}
\examples{
imu_framecombine()
}
