% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imu_proc.R
\name{imu_proc}
\alias{imu_proc}
\title{IMU process}
\usage{
imu_proc(imu.datafile, FOVAngle, GroundLevel = 0, minAlt_dem_atminIMU,
  degree = TRUE, coords.epsg, dem_rast, YawCorrFactor = 0,
  PitchCorrFactor = 0, RollCorrFactor = 0)
}
\description{
This function calculates the corrections for imagery location given the roll, yaw, pitch, and elevation
}
