% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_fun.R
\name{ortho_fun}
\alias{ortho_fun}
\title{orthorectification of imagery}
\usage{
ortho_fun(filenumber, ProcessedIMU, bandtowave, output = c("dataframe",
  "spatial", "none"), printtofile = T)
}
\arguments{
\item{ProcessedIMU}{output from ortho fun}

\item{framex}{frame of interest}

\item{sdf}{spectral data frame output from ortho fun}
}
\description{
This function is the big guy. wraps byframe_corr and spacing_fun, orthorectifies an extensionless bil file that is output by a headwall nano sensor and outputs a dataframe, a sf file, or writes an sf file to folder
}
