% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_crop.R
\name{shp_crop}
\alias{shp_crop}
\title{function to crop data to shapefiles from experiment with distinct plots}
\usage{
shp_crop(proc_img = specdfOUT, shpfile)
}
\description{
function to crop data to shapefiles from experiment with distinct plots
}
