% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shpfile_plotloop.R
\name{shpfile_plotloop}
\alias{shpfile_plotloop}
\title{loops through plots of shapefile to extract spatial hyperspectral data files for each plot}
\usage{
shpfile_plotloop(plotnum, specdfOUT_sf, shpfile, filenumber,
  outputlocation)
}
\description{
for use in shp_crop
}
