% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacing_fun.R
\name{spacing_fun}
\alias{spacing_fun}
\title{spacing function}
\usage{
spacing_fun(pix_i, matchimu, specdfframe)
}
\arguments{
\item{pix_i}{pixel of interest}

\item{matchimu}{output from ortho fun}

\item{specdfframe}{output from ortho fun}
}
\description{
This function orthorectifies one band -- taking the measurements of the drone location, it places a strip of pixels in the correct location on the ground
}
