% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_fun.R
\name{ortho_fun}
\alias{ortho_fun}
\title{orthorectification of imagery}
\usage{
ortho_fun(
  filenumber,
  ProcessedIMU,
  inputlocation,
  outputlocation,
  bandtowave = "standard",
  output = c("dataframe", "spatial", "none"),
  printtospatial = F,
  printtocsv = T
)
}
\arguments{
\item{filenumber}{number of the file to process}

\item{ProcessedIMU}{IMU file created through imu_framecombine, imu_demcombine, and imu_proc.}

\item{inputlocation}{where are the frame files located?}

\item{outputlocation}{where do you want to processed files to go?}

\item{bandtowave}{"standard" uses normal headwall nano values, else call in dataframe that translates band number (1-272) to wavelength in nm (398-1000)}

\item{output}{should function return a non-spatial dataframe, a spatial (sf) file, or nothing?}

\item{printtospatial}{should the function write a spatial file to the processed files folder, defaults to F}

\item{printtocsv}{should the function write a csv file to the processed files folder, defaults to T}
}
\description{
This function is the big guy. wraps byframe_corr, orthorectifies an extensionless bil file that is output by a headwall nano sensor and outputs a dataframe, a sf file, or writes an sf file to folder
}
\examples{
ortho_fun()
}
\keyword{UAV,}
\keyword{broom}
\keyword{ecological}
\keyword{hyperspectral,}
\keyword{orthorectification,}
\keyword{push}
\keyword{research}
\keyword{sensor,}
