% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_crop.R
\name{shp_crop}
\alias{shp_crop}
\title{function to crop data to shapefiles from experiment with distinct plots}
\usage{
shp_crop(proc_img = specdfOUT, shpfile, individualplotstofile = T)
}
\arguments{
\item{filenumber}{number of the file to process}

\item{ProcessedIMU}{IMU file created through imu_framecombine, imu_demcombine, and imu_proc.}

\item{inputlocation}{where are the frame files located?}

\item{outputlocation}{where do you want to processed files to go?}

\item{bandtowave}{"standard" uses normal headwall nano values, else call in dataframe that translates band number (1-272) to wavelength in nm (398-1000)}

\item{output}{should function return a non-spatial dataframe, a spatial (sf) file, or nothing?}

\item{printtofile}{should the function write a spatial file to the processed files folder}
}
\description{
This function merges experiment shape files with the newly created hyperspectral shapefile to crop plots out and/or take plot averages
}
\keyword{UAV,}
\keyword{broom}
\keyword{ecological}
\keyword{hyperspectral,}
\keyword{orthorectification,}
\keyword{push}
\keyword{research}
\keyword{sensor,}
