% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp_crop.R
\name{shp_crop}
\alias{shp_crop}
\title{function to crop data to shapefiles from experiment with distinct plots}
\usage{
shp_crop(proc_img = specdfOUT, shpfile, individualplotstofile = T)
}
\arguments{
\item{proc_img}{processed hyperspectral data, can be df, spatial dataframe, or file name and location}

\item{shpfile}{shapefile of experiment with plots}

\item{individualplotstofile}{do you want to make each plot an individual shapefile?}
}
\description{
This function merges experiment shape files with the newly created hyperspectral shapefile to crop plots out and/or take plot averages
}
\examples{
shp_crop()
}
